requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
keep-alive==0.0.1
MarkupSafe==2.1.5
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
Werkzeug==3.0.4
yarl==1.11.1